compressed_float fused into an FMA on arm64 and wrote different bytes
SerializeCompressedFloat32 computed the quantized integer as one expression:

    uint32(math.Floor(float64(normalizedValue*float32(maxIntegerValue) + 0.5)))

Go permits fusing a multiply and an add into a single FMA unless a conversion
forces the intermediate rounding, and the arm64 backend takes that permission --
it compiles to one FMADDS and rounds ONCE where the format rounds twice.

Measured: value 0.005 over [0, 10] at resolution 0.01 quantized to 0 here where
C++, C#, Rust and the fixed C port all write 1. amd64 emits MULSS+ADDSS, rounds
twice and agrees, so CI on ubuntu-latest stayed green while every Apple Silicon
build emitted different bytes.

The fix is the explicit float32() around the product, which forces the rounding
Go would otherwise be free to skip. STANDARD.md now states the requirement
normatively -- float32, two roundings, no widening, no contraction -- rather
than leaving the precision unspecified, which is what let five implementations
drift apart.

Worth recording why no test caught it: every compressed float value pinned
anywhere in this family lands exactly on a quantum, and an on-quantum value
agrees under float32, double and FMA alike.

+++ write_stream.go
@@ -348,7 +348,17 @@ func (s *WriteStream) SerializeCompressedFloat32(value *float32, min, max, resol
 	}
 	// the clamp bounds the integer value to [0, maxIntegerValue], so it is already
 	// masked to bits
-	integerValue := uint32(math.Floor(float64(normalizedValue*float32(maxIntegerValue) + 0.5)))
+	//
+	// The inner float32() around the product is LOAD BEARING. STANDARD.md
+	// requires float32 arithmetic with TWO roundings — the product rounds
+	// before 0.5 is added — and Go permits fusing a multiply and an add into a
+	// single FMA unless a conversion forces the intermediate rounding. arm64
+	// takes that permission: without this conversion the line compiles to one
+	// FMADDS, rounds once, and writes 0 where every other runtime writes 1 for
+	// value 0.005 over [0, 10] at resolution 0.01. amd64 emits MULSS+ADDSS and
+	// agrees, so CI on x86 stays green while Apple Silicon emits different
+	// bytes. Do not "simplify" this.
+	integerValue := uint32(math.Floor(float64(float32(normalizedValue*float32(maxIntegerValue)) + 0.5)))
 	return s.writeBits(integerValue, bits)
 }
